register rule: Allow region picker to be unset in rule spec.
Once the regions have passed through the rule filter, the operator
may want all resulting regions to be registered to. In the current
situation, the operator would need to write a rule picker
expression which made this possible.

This change allows the rule specification to be created without
specifying a region picker. In this case, all regions that passed
through the filter would immediately move onto the registration
planning.