state: Add new state backend which writes to the local filesystem.
The new file state backend uses the local filesystem to persist
objects. It uses a global lock to avoid data races and writes the
data in JSON format which can be easily viewed by humans.

The in-memory backend is now not enabled by default and the
operator must make a choice on state when running the server. This
adds a little overhead, but it can be as little as a single flag.